+++ src/i18n/i18n.tsx
@@ -15,7 +15,7 @@ const resources = {
 }
 
 i18n
-	.use(initReactI18next as any)
+	.use(initReactI18next)
 	.use(LanguageDetector)
 	.init({
 		resources,