+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- React-scripts (immer) vulnerability
+- Excel file without filename
+
 ## [1.4.0] - 01/03/2021
 
 ### Added